change car_hp to float

+++ common/Dto/player.cpp
@@ -2,5 +2,5 @@
 
 #include "../common_codes.h"
 
-PlayerDto::PlayerDto(const std::string& username, VehicleTipe Type, int  car_hp):
+PlayerDto::PlayerDto(const std::string& username, VehicleTipe Type, float car_hp):
         Dto(uint8_t(ActionCode::SEND_PLAYER)), username(username), Type(Type), car_hp(car_hp) {}
