+++ common/Dto/vehicle_collision.h
@@ -8,12 +8,12 @@
 class VehicleCollisionDto: public Dto {
 public:
     std::string vehicle1_username;
-    int vehicle_1_new_hp; 
+    float vehicle_1_new_hp; 
     std::string vehicle2_username;
-    int vehicle_2_new_hp;
+    float vehicle_2_new_hp;
 
-    VehicleCollisionDto(const std::string& vehicle1_username, int vehicle_1_new_hp, 
-                        const std::string& vehicle2_username, int vehicle_2_new_hp);
+    VehicleCollisionDto(const std::string& vehicle1_username, float vehicle_1_new_hp, 
+                        const std::string& vehicle2_username, float vehicle_2_new_hp);
 };
 
 #endif
